+++ RELEASE_NOTES.md
@@ -3,3 +3,4 @@
 ### Improvements
 
 - Telemetry opt-out option is now hidden behind a Konami code easter egg (up up down down left right left right on Preferences tab)
+- All notification and resource apps are now toggled off by default (users opt-in to which apps they want managed)